api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "neo4j>=5.28.2",
    "neo4j-rust-ext>=5.28.2.1",
    "strawberry-graphql[fastapi]>=0.281.0",
    "typing-extensions>=4.15.0",
]
